# satellite/metabase: revert non-recursive listing optimization